requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.4
requests==2.32.3